requirements.txt
# Reflection Agent Dependencies
# Install with: pip3 install -r requirements.txt

# Core LangChain packages
langchain-core>=0.3.0
langchain-community>=0.3.0

# Google Gemini integration (requires free API key from https://makersuite.google.com/app/apikey)
langchain-google-genai>=3.0.0

# Graph workflow framework
langgraph>=0.2.0

# Data validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Search tool dependencies (for DuckDuckGo)
duckduckgo-search>=6.0.0
